Added email notification feature
Whenever a machine has to be assigned to a technician to fix it, he will receive an email notification with the information about it.

Note: The feature has been added in the 'assign.ejs' file, but it is commented for now to avoid sending massive email notifications when testing the implementation. If we actually want it working we just have to uncomment it in the 'script' at the bottom.